feat: Add script to start arRPC in the background
This commit adds a new script, arrpc_start.ahk, which starts arRPC in the background to autostart it. The script is written in AutoHotkey and requires AutoHotkey version 2.0 or higher. It also disables the tray icon for the script. The script runs the command "node src" in the "C:\Users\Maicol\arrpc" directory and hides the command prompt window.

Author: Maicol Battistini
Date: 2024/10/01
Version: 1.0.0